Create the new `gix-worktree-stream` crate from what was `gix-archive`.
Also, restore `gix-archive` to the previous state.
