--- gix-archive/tests/fixtures/generated-archives/basic.tar.xz
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:939077297db2dbe9917633082205c063bcd35e890b2877d530f2e0f4a75d7e3f
-size 11544
