+++ gix-worktree-stream/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "gix-worktree-stream"
+version = "0.2.0"
+repository = "https://github.com/Byron/gitoxide"
+license = "MIT/Apache-2.0"
+description = "generate a byte-stream from a git-tree"
+authors = ["Sebastian Thiel <sebastian.thiel@icloud.com>"]
+edition = "2021"
+rust-version = "1.65"
+
+[lib]
+doctest = false
+
+[dependencies]
+gix-features = { version = "^0.32.0", path = "../gix-features", features = ["progress", "io-pipe"] }
+gix-hash = { version = "^0.11.3", path = "../gix-hash" }
+gix-object = { version = "^0.33.0", path = "../gix-object" }
+gix-attributes = { version = "^0.15.0", path = "../gix-attributes" }
+gix-filter = { version = "^0.1.0", path = "../gix-filter" }
+gix-traverse = { version = "^0.30.0", path = "../gix-traverse" }
+gix-fs = { version = "^0.4.0", path = "../gix-fs" }
+gix-path = { version = "^0.8.3", path = "../gix-path" }
+
+thiserror = "1.0.26"
+parking_lot = "0.12.1"
+
+[dev-dependencies]
+gix-testtools = { path = "../tests/tools"}
+gix-odb = { path = "../gix-odb"}
+gix-worktree = { path = "../gix-worktree"}
