+++ gix-worktree-stream/tests/fixtures/generated-archives/basic.tar.xz
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:96d22b17f2347b53f29dd5de8b7c217fde40d0289f1b7af8bbeee5c6dd0002fe
+size 11908
